split sift-cli config tests into submodule
