+++ rust/crates/sift_cli/src/cmd/config/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use crate::BIN_NAME;
 use anyhow::{Context, Result, anyhow};
 use crossterm::style::Stylize;
@@ -207,45 +210,4 @@ fn is_update_empty(args: &ConfigUpdateArgs) -> bool {
     grpc_uri.as_ref().is_none_or(|s| s.is_empty())
         && rest_uri.as_ref().is_none_or(|s| s.is_empty())
         && api_key.as_ref().is_none_or(|s| s.is_empty())
-}
-
-#[cfg(test)]
-mod test_is_update_empty {
-    use super::is_update_empty;
-    use crate::cli::ConfigUpdateArgs;
-
-    fn args(
-        grpc_uri: Option<&str>,
-        rest_uri: Option<&str>,
-        api_key: Option<&str>,
-    ) -> ConfigUpdateArgs {
-        ConfigUpdateArgs {
-            interactive: false,
-            grpc_uri: grpc_uri.map(String::from),
-            rest_uri: rest_uri.map(String::from),
-            api_key: api_key.map(String::from),
-        }
-    }
-
-    #[test]
-    fn no_flags_is_empty() {
-        assert!(is_update_empty(&args(None, None, None)));
-    }
-
-    #[test]
-    fn all_empty_strings_is_empty() {
-        assert!(is_update_empty(&args(Some(""), Some(""), Some(""))));
-    }
-
-    #[test]
-    fn any_single_flag_is_not_empty() {
-        assert!(!is_update_empty(&args(Some("g"), None, None)));
-        assert!(!is_update_empty(&args(None, Some("r"), None)));
-        assert!(!is_update_empty(&args(None, None, Some("k"))));
-    }
-
-    #[test]
-    fn all_flags_set_is_not_empty() {
-        assert!(!is_update_empty(&args(Some("g"), Some("r"), Some("k"))));
-    }
 }
